requirements.txt
streamlit>=1.25.0
pdfplumber>=0.10.2
python-dotenv>=1.0.0
requests>=2.31.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
numpy>=1.23.0
